# api-docs-OpenApi-Swagger.json
{"openapi":"3.0.1","info":{"title":"Lemonair API REST","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/lemonair/vuelo/{id}":{"get":{"tags":["Vuelos Info"],"summary":"Consultar vuelo","description":"Devolverá los datos del vuelo identificado por su Id","operationId":"getVueloById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["Editar vuelos"],"summary":"Editar vuelo","description":"Editará el vuelo identificado por su Id con los nuevos datos","operationId":"updateVuelos","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"description":"Introducir Fecha de Despegue en formato UNIX (ej: 1679294143)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VueloDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Editar vuelos"],"summary":"Borrar vuelo","description":"Eliminará el vuelo identificado por su Id de la la lista de vuelos pendientes de despegar","operationId":"deleteVuelos","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"description":"Posibilidad solicitada por cliente pero sin implementación, incluir solo Id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VueloDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]}}}}}}},"/lemonair/salida/{id}/despegue":{"put":{"tags":["Editar vuelos"],"summary":"Vuelo ha despegado","description":"Editará el vuelo identificado por su Id indicando que ha salido del aeropuerto","operationId":"updateVueloStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/lemonair/vuelo":{"get":{"tags":["Vuelos Info"],"summary":"Obtener información de los vuelos programados","description":"Devolverá una lista con los vuelos pendientes de despegar del aeropuerto","operationId":"findByDespegado","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VueloDto"}}}}}}},"post":{"tags":["Editar vuelos"],"summary":"Añadir un nuevo vuelo","description":"Añadirá un nuevo vuelo a la lista de pendientes de despegar del aeropuerto","operationId":"createVuelos","requestBody":{"description":"Introducir Fecha de Despegue en formato UNIX (ej: 1679294143)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VueloDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VueloDto"}}}}}}},"/lemonair/salida":{"get":{"tags":["Aviones Despegados"],"summary":"Devolverá una lista con los aviones que han salido ya del aeropuerto","description":"Identificador, modelo y capacidad de pasajeros","operationId":"findByDespegado_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AvionDto"}}}}}}}},"/lemonair/salida/{id}":{"get":{"tags":["Vuelos Info"],"summary":"Consultar estado del vuelo","description":"Indicará si el vuelo ha salido ya o no y la hora de dicha salida","operationId":"getDespegadoStatusById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/lemonair/info":{"get":{"tags":["Aerolínea Index"],"summary":"Devolverá la información de la aerolínea","description":"Identificador, nombre y número de aviones en la flota","operationId":"getAerolinea","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AerolineaDto"}}}}}}}},"components":{"schemas":{"VueloDto":{"type":"object","properties":{"id":{"type":"integer","description":"id","format":"int64","readOnly":true},"despegado":{"type":"boolean","description":"despegado"},"fechaDespegue":{"type":"string","description":"fechaDespegue","format":"date-time"},"ultimaModificacion":{"type":"string","description":"ultimaModificacion","format":"date-time","readOnly":true},"avionId":{"type":"integer","description":"avionId","format":"int64"}},"description":"Datos del vuelo"},"AvionDto":{"type":"object","properties":{"id":{"type":"integer","description":"Id","format":"int64"},"modelo":{"type":"string","description":"modelo"},"capacidadPasajeros":{"type":"integer","description":"Capacidad de pasajeros","format":"int32"}},"description":"flota de aviones"},"AerolineaDto":{"type":"object","properties":{"id":{"type":"integer","description":"Id","format":"int64"},"nombre":{"type":"string","description":"Nombre"},"sumaAviones":{"type":"integer","description":"Número total de aviones","format":"int32"}},"description":"Datos de la aerolinea"}}}}